Modified CMakeLists to add flags and fixed code flagging warnings.
- Added /W4 (Level 4 warnings) and /WX (Warnings as errors). 
- Fixed failing tests/code issues flagged by these changes in VS2019/VS2022

+++ exercises/practice/armstrong-numbers/.meta/example.cpp
@@ -4,11 +4,11 @@
 
 bool armstrong_numbers::is_armstrong_number(int number) {
     int length = number != 0 && number != 1
-                     ? std::log10(std::abs(static_cast<double>(number))) + 1
+                     ? static_cast<int>(std::log10(std::abs(static_cast<double>(number))) + 1)
                      : 1;
     int accumulate = 0;
     for (int current = number; current != 0; current /= 10) {
-        accumulate += std::pow(current % 10, length);
+        accumulate += static_cast<int>(std::pow(current % 10, length));
     }
     return accumulate == number;
 }
